// web/src/api/generated/model/paymentToConfirm.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { PayMethod } from './payMethod';

export interface PaymentToConfirm {
  id: string;
  reference: string;
  name: string;
  method: PayMethod;
  amount: string;
  created_at: string;
  proof_document_id: string | null;
  expected_by: string;
}